Term upper bound to float and removed parameter in query processing

+++ src/main/java/it/unipi/dii/aide/mircv/InformationRetrievalProject/Indexing/Index.java
@@ -306,7 +306,7 @@ public class Index {
             }
             //we conclude the lexicon merging adding the global posting list length and the term upper bound information.
             fileManager.writeLineOnFile((TextWriter) fileManager.getMyWriterLexicon(), postingListLength + " "
-                    + ((float) Integer.parseInt(terms[0][6])/postingListLength) + "\n");
+                    + (Float.parseFloat(terms[0][6])/postingListLength) + "\n");
         }
         fileManager.closeMergeFiles();
         fileManager.closeScanners();
